third_party/WebKit/LayoutTests/editing/selection/extend/extend_selection_character_01.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script><script>
function testIt(dir, offset, direction, expectedText) {
  test(() => assert_selection(
  `<span dir="${dir}">${makeSample(offset)}</span>`,
  selection => selection.modify('extend', direction, 'character'),
  `<span dir="${dir}">${expectedText}</span>`),
  `${dir} ${offset} extend ${direction} character`
  );
}

function makeSample(offset){
  const kSample = '\nabc \u05D0\u05D1\u05D2 xyz\n';
  return kSample.substr(0, offset) + '|' + kSample.substring(offset);
}

// ltr right character
testIt('ltr', 0, 'right', '\n^a|bc \u05D0\u05D1\u05D2 xyz\n');
testIt('ltr', 1, 'right', '\n^a|bc \u05D0\u05D1\u05D2 xyz\n');
testIt('ltr', 2, 'right', '\na^b|c \u05D0\u05D1\u05D2 xyz\n');
testIt('ltr', 3, 'right', '\nab^c| \u05D0\u05D1\u05D2 xyz\n');
testIt('ltr', 4, 'right', '\nabc^ |\u05D0\u05D1\u05D2 xyz\n');
testIt('ltr', 5, 'right', '\nabc ^\u05D0|\u05D1\u05D2 xyz\n');
testIt('ltr', 6, 'right', '\nabc \u05D0^\u05D1|\u05D2 xyz\n');
testIt('ltr', 7, 'right', '\nabc \u05D0\u05D1^\u05D2| xyz\n');
testIt('ltr', 8, 'right', '\nabc \u05D0\u05D1\u05D2^ |xyz\n');
testIt('ltr', 9, 'right', '\nabc \u05D0\u05D1\u05D2 ^x|yz\n');
testIt('ltr', 10, 'right', '\nabc \u05D0\u05D1\u05D2 x^y|z\n');
testIt('ltr', 11, 'right', '\nabc \u05D0\u05D1\u05D2 xy^z|\n');
testIt('ltr', 12, 'right', '\nabc \u05D0\u05D1\u05D2 xyz|\n');

// ltr left character
testIt('ltr', 0, 'left', '|\nabc \u05D0\u05D1\u05D2 xyz\n');
testIt('ltr', 1, 'left', '\n|abc \u05D0\u05D1\u05D2 xyz\n');
testIt('ltr', 2, 'left', '\n|a^bc \u05D0\u05D1\u05D2 xyz\n');
testIt('ltr', 3, 'left', '\na|b^c \u05D0\u05D1\u05D2 xyz\n');
testIt('ltr', 4, 'left', '\nab|c^ \u05D0\u05D1\u05D2 xyz\n');
testIt('ltr', 5, 'left', '\nabc| ^\u05D0\u05D1\u05D2 xyz\n');
testIt('ltr', 6, 'left', '\nabc |\u05D0^\u05D1\u05D2 xyz\n');
testIt('ltr', 7, 'left', '\nabc \u05D0|\u05D1^\u05D2 xyz\n');
testIt('ltr', 8, 'left', '\nabc \u05D0\u05D1|\u05D2^ xyz\n');
testIt('ltr', 9, 'left', '\nabc \u05D0\u05D1\u05D2| ^xyz\n');
testIt('ltr', 10, 'left', '\nabc \u05D0\u05D1\u05D2 |x^yz\n');
testIt('ltr', 11, 'left', '\nabc \u05D0\u05D1\u05D2 x|y^z\n');
testIt('ltr', 12, 'left', '\nabc \u05D0\u05D1\u05D2 xy|z^\n');

// ltr forward character
testIt('ltr', 0, 'forward', '\n^a|bc \u05D0\u05D1\u05D2 xyz\n');
testIt('ltr', 1, 'forward', '\n^a|bc \u05D0\u05D1\u05D2 xyz\n');
testIt('ltr', 2, 'forward', '\na^b|c \u05D0\u05D1\u05D2 xyz\n');
testIt('ltr', 3, 'forward', '\nab^c| \u05D0\u05D1\u05D2 xyz\n');
testIt('ltr', 4, 'forward', '\nabc^ |\u05D0\u05D1\u05D2 xyz\n');
testIt('ltr', 5, 'forward', '\nabc ^\u05D0|\u05D1\u05D2 xyz\n');
testIt('ltr', 6, 'forward', '\nabc \u05D0^\u05D1|\u05D2 xyz\n');
testIt('ltr', 7, 'forward', '\nabc \u05D0\u05D1^\u05D2| xyz\n');
testIt('ltr', 8, 'forward', '\nabc \u05D0\u05D1\u05D2^ |xyz\n');
testIt('ltr', 9, 'forward', '\nabc \u05D0\u05D1\u05D2 ^x|yz\n');
testIt('ltr', 10, 'forward', '\nabc \u05D0\u05D1\u05D2 x^y|z\n');
testIt('ltr', 11, 'forward', '\nabc \u05D0\u05D1\u05D2 xy^z|\n');
testIt('ltr', 12, 'forward', '\nabc \u05D0\u05D1\u05D2 xyz|\n');

// ltr backward character
testIt('ltr', 0, 'backward', '|\nabc \u05D0\u05D1\u05D2 xyz\n');
testIt('ltr', 1, 'backward', '\n|abc \u05D0\u05D1\u05D2 xyz\n');
testIt('ltr', 2, 'backward', '\n|a^bc \u05D0\u05D1\u05D2 xyz\n');
testIt('ltr', 3, 'backward', '\na|b^c \u05D0\u05D1\u05D2 xyz\n');
testIt('ltr', 4, 'backward', '\nab|c^ \u05D0\u05D1\u05D2 xyz\n');
testIt('ltr', 5, 'backward', '\nabc| ^\u05D0\u05D1\u05D2 xyz\n');
testIt('ltr', 6, 'backward', '\nabc |\u05D0^\u05D1\u05D2 xyz\n');
testIt('ltr', 7, 'backward', '\nabc \u05D0|\u05D1^\u05D2 xyz\n');
testIt('ltr', 8, 'backward', '\nabc \u05D0\u05D1|\u05D2^ xyz\n');
testIt('ltr', 9, 'backward', '\nabc \u05D0\u05D1\u05D2| ^xyz\n');
testIt('ltr', 10, 'backward', '\nabc \u05D0\u05D1\u05D2 |x^yz\n');
testIt('ltr', 11, 'backward', '\nabc \u05D0\u05D1\u05D2 x|y^z\n');
testIt('ltr', 12, 'backward', '\nabc \u05D0\u05D1\u05D2 xy|z^\n');

// rtl right character
testIt('rtl', 0, 'right', '\n^a|bc \u05D0\u05D1\u05D2 xyz\n');
testIt('rtl', 1, 'right', '\n^a|bc \u05D0\u05D1\u05D2 xyz\n');
testIt('rtl', 2, 'right', '\na^b|c \u05D0\u05D1\u05D2 xyz\n');
testIt('rtl', 3, 'right', '\nab^c| \u05D0\u05D1\u05D2 xyz\n');
testIt('rtl', 4, 'right', '\nabc^ |\u05D0\u05D1\u05D2 xyz\n');
testIt('rtl', 5, 'right', '\nabc ^\u05D0|\u05D1\u05D2 xyz\n');
testIt('rtl', 6, 'right', '\nabc \u05D0^\u05D1|\u05D2 xyz\n');
testIt('rtl', 7, 'right', '\nabc \u05D0\u05D1^\u05D2| xyz\n');
testIt('rtl', 8, 'right', '\nabc \u05D0\u05D1\u05D2^ |xyz\n');
testIt('rtl', 9, 'right', '\nabc \u05D0\u05D1\u05D2 ^x|yz\n');
testIt('rtl', 10, 'right', '\nabc \u05D0\u05D1\u05D2 x^y|z\n');
testIt('rtl', 11, 'right', '\nabc \u05D0\u05D1\u05D2 xy^z|\n');
testIt('rtl', 12, 'right', '\nabc \u05D0\u05D1\u05D2 xyz|\n');

// rtl left character
testIt('rtl', 0, 'left', '|\nabc \u05D0\u05D1\u05D2 xyz\n');
testIt('rtl', 1, 'left', '\n|abc \u05D0\u05D1\u05D2 xyz\n');
testIt('rtl', 2, 'left', '\n|a^bc \u05D0\u05D1\u05D2 xyz\n');
testIt('rtl', 3, 'left', '\na|b^c \u05D0\u05D1\u05D2 xyz\n');
testIt('rtl', 4, 'left', '\nab|c^ \u05D0\u05D1\u05D2 xyz\n');
testIt('rtl', 5, 'left', '\nabc| ^\u05D0\u05D1\u05D2 xyz\n');
testIt('rtl', 6, 'left', '\nabc |\u05D0^\u05D1\u05D2 xyz\n');
testIt('rtl', 7, 'left', '\nabc \u05D0|\u05D1^\u05D2 xyz\n');
testIt('rtl', 8, 'left', '\nabc \u05D0\u05D1|\u05D2^ xyz\n');
testIt('rtl', 9, 'left', '\nabc \u05D0\u05D1\u05D2| ^xyz\n');
testIt('rtl', 10, 'left', '\nabc \u05D0\u05D1\u05D2 |x^yz\n');
testIt('rtl', 11, 'left', '\nabc \u05D0\u05D1\u05D2 x|y^z\n');
testIt('rtl', 12, 'left', '\nabc \u05D0\u05D1\u05D2 xy|z^\n');

// rtl forward character
testIt('rtl', 0, 'forward', '\n^a|bc \u05D0\u05D1\u05D2 xyz\n');
testIt('rtl', 1, 'forward', '\n^a|bc \u05D0\u05D1\u05D2 xyz\n');
testIt('rtl', 2, 'forward', '\na^b|c \u05D0\u05D1\u05D2 xyz\n');
testIt('rtl', 3, 'forward', '\nab^c| \u05D0\u05D1\u05D2 xyz\n');
testIt('rtl', 4, 'forward', '\nabc^ |\u05D0\u05D1\u05D2 xyz\n');
testIt('rtl', 5, 'forward', '\nabc ^\u05D0|\u05D1\u05D2 xyz\n');
testIt('rtl', 6, 'forward', '\nabc \u05D0^\u05D1|\u05D2 xyz\n');
testIt('rtl', 7, 'forward', '\nabc \u05D0\u05D1^\u05D2| xyz\n');
testIt('rtl', 8, 'forward', '\nabc \u05D0\u05D1\u05D2^ |xyz\n');
testIt('rtl', 9, 'forward', '\nabc \u05D0\u05D1\u05D2 ^x|yz\n');
testIt('rtl', 10, 'forward', '\nabc \u05D0\u05D1\u05D2 x^y|z\n');
testIt('rtl', 11, 'forward', '\nabc \u05D0\u05D1\u05D2 xy^z|\n');
testIt('rtl', 12, 'forward', '\nabc \u05D0\u05D1\u05D2 xyz|\n');

// rtl backward character
testIt('rtl', 0, 'backward', '|\nabc \u05D0\u05D1\u05D2 xyz\n');
testIt('rtl', 1, 'backward', '\n|abc \u05D0\u05D1\u05D2 xyz\n');
testIt('rtl', 2, 'backward', '\n|a^bc \u05D0\u05D1\u05D2 xyz\n');
testIt('rtl', 3, 'backward', '\na|b^c \u05D0\u05D1\u05D2 xyz\n');
testIt('rtl', 4, 'backward', '\nab|c^ \u05D0\u05D1\u05D2 xyz\n');
testIt('rtl', 5, 'backward', '\nabc| ^\u05D0\u05D1\u05D2 xyz\n');
testIt('rtl', 6, 'backward', '\nabc |\u05D0^\u05D1\u05D2 xyz\n');
testIt('rtl', 7, 'backward', '\nabc \u05D0|\u05D1^\u05D2 xyz\n');
testIt('rtl', 8, 'backward', '\nabc \u05D0\u05D1|\u05D2^ xyz\n');
testIt('rtl', 9, 'backward', '\nabc \u05D0\u05D1\u05D2| ^xyz\n');
testIt('rtl', 10, 'backward', '\nabc \u05D0\u05D1\u05D2 |x^yz\n');
testIt('rtl', 11, 'backward', '\nabc \u05D0\u05D1\u05D2 x|y^z\n');
testIt('rtl', 12, 'backward', '\nabc \u05D0\u05D1\u05D2 xy|z^\n');
</script>
